Add files via upload
I have added gif popouts, and I can't find the bug on the function that presents the loser images.  The script is showing on the right answer response.  I will continue to work on it, but I wanted to turn in what I have built line by line with the tutor and my resources.

I am also planning to pull in the ajax and Api work that we are working on.